Extract Analytics ID field into separate class Move OPTION constants into corrsponding classes.
