+++ site-performance-tracker.php
@@ -59,6 +59,7 @@ if ( version_compare( phpversion(), '5.3', '<' ) ) {
 require_once __DIR__ . '/php/src/class-plugin.php';
 require_once __DIR__ . '/php/src/class-settings.php';
 require_once __DIR__ . '/php/src/class-analyticstypesfield.php';
+require_once __DIR__ . '/php/src/class-analyticsidfield.php';
 require_once __DIR__ . '/php/helpers.php';
 
 /**
